I think I got all the variables this time

+++ src/centroiders.cpp
@@ -29,8 +29,8 @@ std::vector<Star> DummyCentroidAlgorithm::Go(unsigned char *image, int imageWidt
 // a poorly designed thresholding algorithm
 int BadThreshold(unsigned char *image, int imageWidth, int imageHeight) {
     //loop through entire array, find sum of magnitudes
-    int totalMag = 0;
-    for (int i = 0; i < imageHeight * imageWidth; i++) {
+    long totalMag = 0;
+    for (long i = 0; i < imageHeight * imageWidth; i++) {
         totalMag += image[i];
     }
     return (((totalMag/(imageHeight * imageWidth)) + 1) * 15) / 10;
@@ -51,7 +51,7 @@ int OtsusThreshold(unsigned char *image, int imageWidth, int imageHeight) {
 
     memset(histogram, 0, sizeof(int)*256);
 
-    for (int i = 0; i < total; i++) {
+    for (long i = 0; i < total; i++) {
         histogram[image[i]] ++;
     }
     for (int i = 0; i < 256; i ++) {
@@ -81,11 +81,11 @@ int BasicThreshold(unsigned char *image, int imageWidth, int imageHeight) {
     unsigned long totalMag = 0;
     float std = 0;
     long totalPixels = imageHeight * imageWidth;
-    for (int i = 0; i < totalPixels; i++) {
+    for (long i = 0; i < totalPixels; i++) {
         totalMag += image[i];
     }
     float mean = totalMag / totalPixels;
-    for (int i = 0; i < totalPixels; i++) {
+    for (long i = 0; i < totalPixels; i++) {
         std += std::pow(image[i] - mean, 2);
     }
     std = std::sqrt(std / totalPixels);
@@ -98,7 +98,7 @@ int BasicThresholdOnePass(unsigned char *image, int imageWidth, int imageHeight)
     float std = 0;
     float sq_totalMag = 0;
     long totalPixels = imageHeight * imageWidth;
-    for (int i = 0; i < totalPixels; i++) {
+    for (long i = 0; i < totalPixels; i++) {
         totalMag += image[i];
         sq_totalMag += image[i] * image[i];
     }
@@ -111,7 +111,7 @@ int BasicThresholdOnePass(unsigned char *image, int imageWidth, int imageHeight)
 struct CentroidParams {
     float yCoordMagSum; 
     float xCoordMagSum;
-    int magSum;
+    long magSum;
     int xMin;
     int xMax;
     int yMin;
@@ -121,7 +121,7 @@ struct CentroidParams {
 };
 
 //recursive helper here
-void CogHelper(CentroidParams &p, int i, unsigned char *image, int imageWidth, int imageHeight) {
+void CogHelper(CentroidParams &p, long i, unsigned char *image, int imageWidth, int imageHeight) {
     if (i >= 0 && i < imageWidth * imageHeight && image[i] >= p.cutoff && p.checkedIndices.count(i) == 0) {
         p.checkedIndices.insert(i);
         if (i % imageWidth > p.xMax) {
@@ -154,7 +154,7 @@ std::vector<Star> CenterOfGravityAlgorithm::Go(unsigned char *image, int imageWi
     std::vector<Star> result;
 
     p.cutoff = BasicThreshold(image, imageWidth, imageHeight);
-    for (int i = 0; i < imageHeight * imageWidth; i++) {
+    for (long i = 0; i < imageHeight * imageWidth; i++) {
         if (image[i] >= p.cutoff && p.checkedIndices.count(i) == 0) {
 
             //iterate over pixels that are part of the star
@@ -198,7 +198,7 @@ struct IWCoGParams {
     std::unordered_set<int> checkedIndices;
 };
 
-void IWCoGHelper(IWCoGParams &p, int i, unsigned char *image, int imageWidth, int imageHeight, std::vector<int> &starIndices) {
+void IWCoGHelper(IWCoGParams &p, long i, unsigned char *image, int imageWidth, int imageHeight, std::vector<int> &starIndices) {
     if (i >= 0 && i < imageWidth * imageHeight && image[i] >= p.cutoff && p.checkedIndices.count(i) == 0) {
         p.checkedIndices.insert(i);
         starIndices.push_back(i);
@@ -231,7 +231,7 @@ Stars IterativeWeightedCenterOfGravityAlgorithm::Go(unsigned char *image, int im
     IWCoGParams p;
     std::vector<Star> result;
     p.cutoff = BasicThreshold(image, imageWidth, imageHeight);
-    for (int i = 0; i < imageHeight * imageWidth; i++) {
+    for (long i = 0; i < imageHeight * imageWidth; i++) {
         //check if pixel is part of a "star" and has not been iterated over
         if (image[i] >= p.cutoff && p.checkedIndices.count(i) == 0) {
             std::vector<int> starIndices; //indices of the current star
@@ -277,7 +277,7 @@ Stars IterativeWeightedCenterOfGravityAlgorithm::Go(unsigned char *image, int im
                 xWeightedCoordMagSum = 0;
                 weightedMagSum = 0;
                 stop++;
-                for (int j = 0; j < (int) starIndices.size(); j++) {
+                for (long j = 0; j < starIndices.size(); j++) {
                     //calculate w
                     float currXCoord = (float) (starIndices.at(j) % imageWidth);
                     float currYCoord = (float) (starIndices.at(j) / imageWidth);
